refactor: split layout, widgets, and background into sub-modules
- layout/mod.rs (1184→367 LOC): extract compute helpers to layout/compute.rs
- widgets.rs (966→3 files): split into widgets/mod.rs + hud.rs + tooltip.rs
- background.rs (946→3 files): split into background/mod.rs + time_of_day.rs + lighting.rs

Pure code organization — no behavior changes. All 294 tests pass.
